+++ Assets/uNode3/Core.Editor/Window/AboutWindow.cs
@@ -3,8 +3,8 @@ using UnityEditor;
 
 namespace MaxyGames.UNode.Editors {
 	public class About : EditorWindow {
-		public const string version = "3.0.6";
-		public const float compilerVersion = 3.06f;
+		public const string version = "3.0.7";
+		public const float compilerVersion = 3.07f;
 
 		GUIStyle PublisherNameStyle, headerStyle, infoStyle;
 		GUIStyle ToolBarStyle;