+++ Assets/uNode3/Core.Editor/UIElementGraph/Views/NodeViews/BaseNodeView.cs
@@ -209,6 +209,7 @@ namespace MaxyGames.UNode.Editors {
 			if(ShowExpandButton()) {
 				expanded = nodeObject.nodeExpanded;
 			}
+			UpdateUI();
 		}
 
 		/// <summary>