+++ Assets/uNode3/changelog.md
@@ -3,7 +3,9 @@
 ## v3.0.7
 - Added default value support for parameters
 - Added Graph Inspector window, useful if needed to see both node properties and game object properties on the screen in the same time
+- Update Roslyn to latest version
 - Improved c# parser
+- Improved Editor
 - Fixed some bugs
 
 ## v3.0.6